Module webb_proposal_signing_backends::queue

A module to handle the queue of proposals

Modules

mem

A module for in-memory Proposals Queue.

policy

A module for Proposals Polices.

Structs

QueuedAnchorUpdateProposal

A Small Wrapper around a proposal that adds some metadata.

QueuedProposalMetadata

Associated metadata for a queued proposal.

Traits

ProposalHash

A small trait to add more functionality to the [ProposalTrait]. This trait is used to calculate the hash of the proposal.

ProposalMetadata

The ProposalMetadata trait is implemented by proposals that carry metadata. Metadata is additional information that accompanies a proposal, like its nonce, creation time, etc.

ProposalsQueue

A Proposal Queue is a simple Queue that holds the proposals that are going to be signed or already signed and need to be sent to the target system to be executed.

Functions

run

Runs the queue in a loop that it will try to dequeue proposals and sends them to the signing backend.