Trait webb_proposal_signing_backends::queue::policy::ProposalPolicy

pub trait ProposalPolicy {
    // Required method
    fn check<Q: ProposalsQueue>(
        &self,
        proposal: &Q::Proposal,
        queue: &Q
    ) -> Result<()>;
}

The ProposalPolicy trait defines the behavior of a policy that is applied to proposal in a queue. A policy determines whether a proposal should be accepted or rejected based on specific criteria.

Required Methods

fn check<Q: ProposalsQueue>( &self, proposal: &Q::Proposal, queue: &Q ) -> Result<()>

Checks whether a proposal should be accepted or rejected based on the policy’s criteria.

Arguments

proposal: A reference to the proposal to be checked.

queue: A reference to the proposals queue that contains the proposal.

Errors

This method may return an error if the proposal fails to meet the policy’s criteria.