Trait webb_proposal_signing_backends::queue::ProposalsQueue

pub trait ProposalsQueue {
    type Proposal: ProposalTrait + ProposalMetadata + Send + Sync + 'static;

    // Required methods
    fn enqueue<Policy: ProposalPolicy>(
        &self,
        proposal: Self::Proposal,
        policy: Policy
    ) -> Result<()>;
    fn dequeue<Policy: ProposalPolicy>(
        &self,
        policy: Policy
    ) -> Result<Option<Self::Proposal>>;
    fn len(&self) -> Result<usize>;
    fn clear(&self) -> Result<()>;
    fn retain<F>(&self, f: F) -> Result<()>
       where F: FnMut(&Self::Proposal) -> bool;
    fn modify_in_place<F>(&self, f: F) -> Result<()>
       where F: FnMut(&mut Self::Proposal) -> Result<()>;
    fn find<F>(&self, f: F) -> Result<Option<Self::Proposal>>
       where F: FnMut(&Self::Proposal) -> bool;

    // Provided method
    fn is_empty(&self) -> Result<bool> { ... }
}

A Proposal Queue is a simple Queue that holds the proposals that are going to be signed or already signed and need to be sent to the target system to be executed.

Required Associated Types

type Proposal: ProposalTrait + ProposalMetadata + Send + Sync + 'static

The proposal type associated with the Proposal Queue.

Required Methods

fn enqueue<Policy: ProposalPolicy>( &self, proposal: Self::Proposal, policy: Policy ) -> Result<()>

Enqueues a proposal into the queue with the specified policy.

The [policy::ProposalsQueuePolicy] trait is implemented for tuples of up to 5 policies. These policies are chained together and applied sequentially. If any of the policies in the chain rejects the proposal, the enqueue operation fails and returns an error. The proposal is added to the queue, and the provided policy is applied to determine if the proposal should be accepted. If the policy chain rejects the proposal, the enqueue operation fails and returns an error.

Arguments

proposal - A reference to the proposal to be enqueued.

policy - The policy to be applied to the proposal. The policy determines if the proposal should be accepted.

Returns

Returns a Result indicating whether the enqueue operation was successful or not. If the enqueue operation fails, an error is returned. Possible errors include:

EnqueueQueueError if there is an issue during the enqueue operation, such as a policy rejection.

fn dequeue<Policy: ProposalPolicy>( &self, policy: Policy ) -> Result<Option<Self::Proposal>>

Dequeues a proposal from the queue with the specified policy.

The dequeue operation retrieves the next proposal from the queue that satisfies the provided policy. The policy is applied sequentially to the proposals in the queue until a matching proposal is found. The [policy::ProposalsQueuePolicy] trait is implemented for tuples of up to 5 policies. These policies are chained together and applied sequentially. If any of the policies in the chain rejects the proposal, the dequeue operation fails and returns an error.

Arguments

policy - The policy to be applied in the dequeue operation. The policy determines if a proposal should be dequeued.

Returns

Returns a Result containing an optional proposal.

If a proposal satisfying the policy is found, it is returned as Some(proposal).

If no matching proposal is found, None is returned. If the dequeue operation fails, an error is returned. Possible errors include:

DequeueQueueError if there is an issue during the dequeue operation.

fn len(&self) -> Result<usize>

Returns the length of the proposal queue.

The length of the queue is the total number of proposals currently stored in it.

Returns

Returns a Result containing the length of the queue.

fn clear(&self) -> Result<()>

Clears the proposal queue, removing all proposals.

This operation removes all proposals from the queue, resulting in an empty queue.

Returns

Returns a Result indicating whether the clear operation was successful or not. If the clear operation fails, an error is returned. Possible errors include:

ClearQueueError if there is an issue during the clear operation.

fn retain<F>(&self, f: F) -> Result<()>where F: FnMut(&Self::Proposal) -> bool,

Retains only the proposals in the queue that satisfy the given predicate.

This method removes all proposals from the queue that do not satisfy the provided predicate function. The predicate function takes a reference to a proposal, and returns true if the proposal should be retained, or false if it should be removed from the queue.

Arguments

f - A mutable closure or function that determines whether a proposal should be retained or removed. The closure takes a reference to a proposal, and returns a boolean value. - If the closure returns true, the proposal is retained in the queue. - If the closure returns false, the proposal is removed from the queue.

Returns

Returns a Result indicating whether the retain operation was successful or not. If the retain operation fails, an error is returned. Possible errors include:

RetainQueueError if there is an issue during the retain operation.

fn modify_in_place<F>(&self, f: F) -> Result<()>where F: FnMut(&mut Self::Proposal) -> Result<()>,

Modifies a proposal in place using the provided closure.

This method allows modifying a proposal in the queue using a mutable closure. The closure takes a mutable reference to a proposal and performs the required modifications.

Arguments

f - A mutable closure or function that modifies the proposal in place. The closure takes a mutable reference to the proposal and returns a Result. - If the closure returns Ok(()), the modification is successful. - If the closure returns Err(_), the modification fails and an error is returned.

Returns

Returns a Result indicating whether the modification operation was successful or not. If the modification operation fails, an error is returned. Possible errors include:

ModifyQueueError if there is an issue during the modification operation.

fn find<F>(&self, f: F) -> Result<Option<Self::Proposal>>where F: FnMut(&Self::Proposal) -> bool,

Finds and returns a proposal that satisfies the given predicate. The predicate function takes a reference to a proposal, and returns true if the proposal is a match, or false if it is not a match. The first proposal that satisfies the predicate is returned. If no matching proposal is found, None is returned.

Provided Methods

fn is_empty(&self) -> Result<bool>

Returns true if the proposal queue is empty.

Returns

Returns a Result containing a boolean value.